# backend/group/internal/rest/controller/leavegroup/handle.go
package leavegroup

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"

	"github.com/FSpruhs/kick-app/backend/group/internal/application"
	"github.com/FSpruhs/kick-app/backend/group/internal/application/commands"
)

// InviteUser godoc
// @Summary      user leaves a group
// @Description  user leaves a group
// @Tags         group
// @Accepted       json
// @Produce      json
// @Success      201
// @Failure      400
// @Failure      500
// @Router       /group/user [delete]
func Handle(app application.App) gin.HandlerFunc {
	return func(c *gin.Context) {
		var message Message

		if err := c.BindJSON(&message); err != nil {
			c.JSON(http.StatusBadRequest, c.Error(err))

			return
		}

		if err := validator.New().Struct(&message); err != nil {
			c.JSON(http.StatusBadRequest, c.Error(err))

			return
		}

		command := commands.LeaveGroup{
			GroupID: message.GroupID,
			UserID:  message.UserID,
		}

		if err := app.LeaveGroup(&command); err != nil {
			c.JSON(http.StatusInternalServerError, c.Error(err))

			return
		}

		c.JSON(http.StatusOK, nil)
	}
}